requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
litellm>=1.40.0
langsmith>=0.1.120
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.30.0
pydantic>=2.0.0
jsonschema>=4.0.0
